+++ crates/adapterlib/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "feldera-adapterlib"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+with-avro = ["apache-avro"]
+
+[dependencies]
+feldera-types = { path = "../feldera-types" }
+dbsp = { path = "../dbsp" }
+anyhow = { version = "1.0.57", features = ["backtrace"] }
+serde_yaml = "0.9.34+deprecated"
+actix-web = { version = "4.4.0", default-features = false, features = ["cookies", "macros", "compress-gzip", "compress-brotli"] }
+erased-serde = "0.3.23"
+dyn-clone = "1.0.17"
+arrow = { version = "52.0.0", features = ["chrono-tz"] }
+apache-avro = { version = "0.17.0", optional = true }
+serde_arrow = { version = "0.11.6", features = ["arrow-52"] }
+serde = { version = "1.0.213", features = ["derive"] }
+serde_json = { version = "1.0.127", features = ["raw_value"] }
+rmp-serde = "1.3.0"
+rmpv = { version = "1.3.0", features = ["with-serde"] }
+bytemuck = "1.16.3"
+num-traits = "0.2.15"
+num-derive = "0.3.3"
+
+[package.metadata.cargo-machete]
+ignored = ["num-traits"]
