+++ crates/adapterlib/README.md
@@ -1,0 +1,4 @@
+# Feldera I/O adapter library
+
+This crate provides base traits and types to implement input and output
+adapters for the Feldera I/O adapter framework.
